Fix zsh shell hook performance issue
The precmd hook is causing a significant overhead on every command
executed in the shell.
The shell hook does not need to be this strict however, it can also
run only on directory changes, ensuring that the shell remains snappy
while executing commands.

Also contains a similar fix for fish.

Fixes #259.